Parameter Tff.T

include Dolmen_intf.Term.Smtlib_Arith_Common with type t := Type.T.t with type cst := Type.T.Const.t

val mk : string -> Type.T.t

Build a constant. The literal is passed as a string to avoid overflow caused by the limited precision of native number formats.

val minus : Type.T.t -> Type.T.t

Arithmetic unary minus/negation.

val add : Type.T.t -> Type.T.t -> Type.T.t

Arithmetic addition.

val sub : Type.T.t -> Type.T.t -> Type.T.t

Arithmetic substraction

val mul : Type.T.t -> Type.T.t -> Type.T.t

Arithmetic multiplication

val lt : Type.T.t -> Type.T.t -> Type.T.t

Arithmetic "less than" comparison.

val le : Type.T.t -> Type.T.t -> Type.T.t

Arithmetic "less or equal" comparison.

val gt : Type.T.t -> Type.T.t -> Type.T.t

Arithmetic "greater than" comparison.

val ge : Type.T.t -> Type.T.t -> Type.T.t

Arithmetic "greater or equal" comparison.

val div' : Type.T.Const.t

Constant for the division.

val div : Type.T.t -> Type.T.t -> Type.T.t

Euclidian division. See Smtlib theory for a full description.

val rem' : Type.T.Const.t

Constant for the remainder.

val rem : Type.T.t -> Type.T.t -> Type.T.t

Euclidian integer remainder See Smtlib theory for a full description.

val abs : Type.T.t -> Type.T.t

Arithmetic absolute value.

val divisible : string -> Type.T.t -> Type.T.t

Arithmetic divisibility predicate. Indexed over constant integers (represented as strings, see int).