Module Real.Tff

Parameters

module Type : Tff_intf.S

module Ty : Dolmen.Intf.Ty.Smtlib_Real with type t := Type.Ty.t

module T : Dolmen.Intf.Term.Smtlib_Real with type t := Type.T.t and type cst := Type.T.Const.t

Signature

type Type.warn +=

| Restriction : config * string -> Dolmen.Std.Term.t Type.warn
(*
Warning for expressions which tecnically do not respect the strict spec but respect the large spec.
*)

Arithmetic type-checking warnings

type Type.err +=

| Forbidden : config * string -> Dolmen.Std.Term.t Type.err
(*
Error for expressions which do not respect the spec.
*)

Arithmetic type-checking errors

val parse : config:config -> Dolmen.Smtlib2.version -> Type.builtin_symbols

Parsing function for type-checking