Module Real_Int.Tff

Parameters

module Type : Tff_intf.S

module Ty : Dolmen.Intf.Ty.Smtlib_Real_Int with type t := Type.Ty.t

module T : Dolmen.Intf.Term.Smtlib_Real_Int with type t := Type.T.t and type ty := Type.Ty.t and type Int.cst := Type.T.Const.t and type Real.cst := Type.T.Const.t

Signature

type Type.warn +=

| Restriction : string -> Dolmen.Std.Term.t Type.warn	(*
Warning for expressions which tecnically do not respect the strict spec but respect the large spec.
*)

Arithmetic type-checking warnings

type Type.err +=

| Forbidden : string -> Dolmen.Std.Term.t Type.err	(*
Error for expressions which do not respect the spec.
*)

| Expected_arith_type : Type.Ty.t -> Dolmen.Std.Term.t Type.err	(*
Error raised when an arithmetic type was expected (i.e. either int or real), but another type was found.
*)

Additional errors specific to arithmetic typing.

val parse : arith:arith -> Dolmen.Smtlib2.version -> Type.builtin_symbols

Parsing function for type-checking