Parameter Tff.2-Ty

include Dolmen.Intf.Ty.Smtlib_Int with type t := Type.Ty.t

val int : Type.Ty.t

The type for integer expressions.

include Dolmen.Intf.Ty.Smtlib_Real with type t := Type.Ty.t

val real : Type.Ty.t

The type for integer expressions.

type view#row

Partial view for types. These are used by the Reals_Ints theory to perform type-based dispatch, and automatic conversion of Ints to Reals when specified by the specification.

and view = private [>

| `Int

| `Real

]

Partial view for types. These are used by the Reals_Ints theory to perform type-based dispatch, and automatic conversion of Ints to Reals when specified by the specification.

val view : Type.Ty.t -> view

Partial view of a type.