Parameter Tff.T

val select : Type.T.t -> Type.T.t -> Type.T.t

select arr idx creates the get operation on functionnal array arr for index idx.

val store : Type.T.t -> Type.T.t -> Type.T.t -> Type.T.t

store arr idx value creates the set operation on functional array arr for value value at index idx.