Parameter Tff.Ty

val array : Type.Ty.t -> Type.Ty.t -> Type.Ty.t

The type of functionnal arrays from one type to another.

type view = private [>

| `Int

| `Real

| `Bitv of int

| `Array of Type.Ty.t * Type.Ty.t

]

Partial views for types. These are used in the Array theory to enforce some restrictions logics impose on the types of arrays that cna occur.

val view : Type.Ty.t -> view

Partial view of a type.