Parameter Tff.3-T

type t

The type of terms

val select : t -> t -> t

select arr idx creates the get operation on functionnal array arr for index idx.

val store : t -> t -> t -> t

store arr idx value creates the set operation on functional array arr for value value at index idx.