Parameter Tff.T

val mk : string -> Type.T.t

Create a bitvector litteral from a string representation. The string should only contain characters '0' or '1'.

val concat : Type.T.t -> Type.T.t -> Type.T.t

Bitvector concatenation.

val extract : int -> int -> Type.T.t -> Type.T.t

Bitvector extraction, using in that order, the start and then end the position of the bitvector to extract.