Parameter Tff.4-T

type t = Type.T.t

The type of terms.

type cstr

The type of ADT constructor

val cstr_tester : cstr -> t -> t

Given a constructor c and a term t, returns a terms that evaluates to true iff t has c as head constructor.

val _true : t

The smybol for true

val _false : t

The symbol for false

val neg : t -> t

Negation.

val _or : t list -> t

Disjunction of formulas

val _and : t list -> t

Disjunction of formulas

val nand : t -> t -> t

Not-and

val nor : t -> t -> t

Not-or

val imply : t -> t -> t

Implication

val equiv : t -> t -> t

Equivalence

val xor : t -> t -> t

Exclusive disjunction.

val ite : t -> t -> t -> t

ite condition then_t else_t creates a conditional branch.

val eq : t -> t -> t

Create a chain of equalities.

val distinct : t list -> t

Distinct constraints on terms.