Parameter Thf.1-Type

module Tag : Dolmen_intf.Tag.S

module Ty : Dolmen_intf.Ty.Thf with type 'a tag := 'a Tag.t and type path := Dolmen_std.Path.t

module T : Dolmen_intf.Term.Thf with type ty := Ty.t and type ty_var := Ty.Var.t and type ty_const := Ty.Const.t and type 'a tag := 'a Tag.t and type path := Dolmen_std.Path.t

include Intf.Formulas with type ty := Ty.t and type ty_var := Ty.Var.t and type ty_cst := Ty.Const.t and type term := T.t and type term_var := T.Var.t and type term_cst := T.Const.t and type term_cstr := T.Cstr.t and type term_field := T.Field.t and type 'a ast_tag := 'a Tag.t

type ty

types

type ty_var

type ty_cst

type term

type term_var

type term_cst

type term_cstr

type term_field

type 'a ast_tag

Type definitions

type order =

| First_order
First-oreder typechecking

| Higher_order
Higher-order typechecking

Control whether the typechecker should type

type poly =

| Explicit
Type arguments must be explicitly given in funciton applications

| Implicit
Type arguments are not given in funciton applications, and instead type annotations/coercions are used to disambiguate applications of polymorphic symbols.

| Flexible
Mix between explicit and implicit: depending on the arity of a symbol and the number of arguments provided, either the provided type arguments are used, or wildcards are generated for all of them, and later instantiated when needed.

The various polymorphism mode for the typechecker

type sym_inference_source = {

symbol : Dolmen_std.Id.t;

symbol_loc : Dolmen_std.Loc.t;

mutable inferred_ty : ty;

}

type var_inference_source = {

variable : Dolmen_std.Id.t;

variable_loc : Dolmen_std.Loc.t;

mutable inferred_ty : ty;

}

type wildcard_source =

| Arg_of of wildcard_source

| Ret_of of wildcard_source

| From_source of Dolmen_std.Term.t

| Added_type_argument of Dolmen_std.Term.t

| Symbol_inference of sym_inference_source

| Variable_inference of var_inference_source

type wildcard_shape =

| Forbidden

| Any_in_scope

| Any_base of {

allowed : ty list;

preferred : ty;

}

| Arrow of {

arg_shape : wildcard_shape;

ret_shape : wildcard_shape;

}

type infer_term_scheme =

| No_inference

| Wildcard of wildcard_shape

type var_infer = {

infer_type_vars : bool;

infer_term_vars : infer_term_scheme;

}

Specification of how to infer variables.

type sym_infer = {

infer_type_csts : bool;

infer_term_csts : infer_term_scheme;

}

Specification of how to infer symbols.

type expect =

| Type

| Term

| Anything

type tag =

| Set : 'a ast_tag * 'a -> tag

| Add : 'a list ast_tag * 'a -> tag
Existencial wrapper around tags

type res =

| Ttype

| Ty of ty

| Term of term

| Tags of tag list

The results of parsing an untyped term.

type builtin_res = [

| `Ttype of Dolmen_std.Term.t -> Dolmen_std.Term.t list -> unit

| `Ty of Dolmen_std.Term.t -> Dolmen_std.Term.t list -> ty

| `Term of Dolmen_std.Term.t -> Dolmen_std.Term.t list -> term

| `Tags of Dolmen_std.Term.t -> Dolmen_std.Term.t list -> tag list

]

The result of parsing a symbol by the theory

type not_found = [

| `Not_found

]

Not bound bindings

type inferred =

| Ty_fun of ty_cst

| Term_fun of term_cst

The type for inferred symbols.

type reason =

| Builtin

| Bound of Dolmen_std.Loc.file * Dolmen_std.Term.t

| Inferred of Dolmen_std.Loc.file * Dolmen_std.Term.t

| Defined of Dolmen_std.Loc.file * Dolmen_std.Statement.def

| Declared of Dolmen_std.Loc.file * Dolmen_std.Statement.decl
The type of reasons for constant typing

type binding = [

| `Not_found

| `Builtin of [ `Ttype | `Ty | `Term | `Tag ]

| `Variable of [ `Ty of ty_var * reason option | `Term of term_var * reason option ]

| `Constant of [ `Ty of ty_cst * reason option | `Cstr of term_cstr * reason option | `Term of term_cst * reason option | `Field of term_field * reason option ]

]

The bindings that can occur.

type nonrec symbol = Intf.symbol =

| Id of Dolmen_std.Id.t

| Builtin of Dolmen_std.Term.builtin

Wrapper around potential function symbols from the Dolmen AST.

Errors and warnings

type _ fragment =

| Ast : Dolmen_std.Term.t -> Dolmen_std.Term.t fragment

| Def : Dolmen_std.Statement.def -> Dolmen_std.Statement.def fragment

| Defs : Dolmen_std.Statement.defs -> Dolmen_std.Statement.defs fragment

| Decl : Dolmen_std.Statement.decl -> Dolmen_std.Statement.decl fragment

| Decls : Dolmen_std.Statement.decls -> Dolmen_std.Statement.decls fragment

| Located : Dolmen_std.Loc.t -> Dolmen_std.Loc.t fragment

Fragments of input that represent the sources of warnings/errors

type _ warn = ..

The type of warnings, parameterized by the type of fragment they can trigger on

type warn += Unused_type_variable : ty_var -> Dolmen_std.Term.t warn | Unused_term_variable : term_var -> Dolmen_std.Term.t warn | Error_in_attribute : exn -> Dolmen_std.Term.t warn | Superfluous_destructor : Dolmen_std.Id.t * Dolmen_std.Id.t * term_cst -> Dolmen_std.Term.t warn

Warnings that cna trigger on regular parsed terms.

type warn += Shadowing : Dolmen_std.Id.t * binding * binding -> _ warn

Special case of warnings for shadowing, as it can happen both from a term but also a declaration, hence why the type variable of warn is left wild.

type _ err = ..

The type of errors, parameterized by the type of fragment they can trigger on

type err += Not_well_founded_datatypes : Dolmen_std.Statement.decl list -> Dolmen_std.Statement.decls err

Errors that occur on declaration(s)

type err += Expected : string * res option -> Dolmen_std.Term.t err | Bad_index_arity : string * int * int -> Dolmen_std.Term.t err | Bad_ty_arity : ty_cst * int -> Dolmen_std.Term.t err | Bad_op_arity : symbol * int list * int -> Dolmen_std.Term.t err | Bad_cstr_arity : term_cstr * int list * int -> Dolmen_std.Term.t err | Bad_term_arity : term_cst * int list * int -> Dolmen_std.Term.t err | Bad_poly_arity : ty_var list * ty list -> Dolmen_std.Term.t err | Over_application : term list -> Dolmen_std.Term.t err | Repeated_record_field : term_field -> Dolmen_std.Term.t err | Missing_record_field : term_field -> Dolmen_std.Term.t err | Mismatch_record_type : term_field * ty_cst -> Dolmen_std.Term.t err | Mismatch_sum_type : term_cstr * ty -> Dolmen_std.Term.t err | Var_application : term_var -> Dolmen_std.Term.t err | Ty_var_application : ty_var -> Dolmen_std.Term.t err | Type_mismatch : term * ty -> Dolmen_std.Term.t err | Quantified_var_inference : Dolmen_std.Term.t err | Unhandled_builtin : Dolmen_std.Term.builtin -> Dolmen_std.Term.t err | Cannot_tag_tag : Dolmen_std.Term.t err | Cannot_tag_ttype : Dolmen_std.Term.t err | Cannot_find : Dolmen_std.Id.t * string -> Dolmen_std.Term.t err | Type_var_in_type_constructor : Dolmen_std.Term.t err | Forbidden_quantifier : Dolmen_std.Term.t err | Missing_destructor : Dolmen_std.Id.t -> Dolmen_std.Term.t err | Type_def_rec : Dolmen_std.Statement.def -> Dolmen_std.Statement.defs err | Higher_order_application : Dolmen_std.Term.t err | Higher_order_type : Dolmen_std.Term.t err | Higher_order_env_in_tff_typechecker : Dolmen_std.Loc.t err | Polymorphic_function_argument : Dolmen_std.Term.t err | Non_prenex_polymorphism : ty -> Dolmen_std.Term.t err | Inference_forbidden : ty_var * wildcard_source * ty -> Dolmen_std.Term.t err | Inference_conflict : ty_var * wildcard_source * ty * ty list -> Dolmen_std.Term.t err | Inference_scope_escape : ty_var * wildcard_source * ty_var * reason option -> Dolmen_std.Term.t err | Unbound_variables : ty_var list * term_var list * term -> Dolmen_std.Term.t err | Uncaught_exn : exn * Stdlib.Printexc.raw_backtrace -> Dolmen_std.Term.t err | Unhandled_ast : Dolmen_std.Term.t err

Errors that occur on regular parsed terms.

Global State

type state

The type of mutable state for typechecking.

val new_state : unit -> state

Create a new state.

val copy_state : state -> state

Make a copy of the global state included in the env

Typing Environment

type env

The type of environments for typechecking.

type 'a typer = env -> Dolmen_std.Term.t -> 'a

A general type for typers. Takes a local environment and the current untyped term, and return a value. The typer may need additional information for parsing, in which case the return value will be a function.

raises Typing_error

type builtin_symbols = env -> symbol -> [ builtin_res | not_found ]

The type of a typer for builtin symbols. Given the environment and a symbol, the theory should return a typing function if the symbol belongs to the theory. This typing function takes first the ast term of the whole application that is beign typechecked, and the list of arguments to the symbol.

type warning =

| Warning : env * 'a fragment * 'a warn -> warning

Existential wrapper around warnings

type error =

| Error : env * 'a fragment * 'a err -> error

Existential wrapper around errors

exception Typing_error of error

Exception for typing errors

val empty_env : ?⁠st:state -> ?⁠expect:expect -> ?⁠var_infer:var_infer -> ?⁠sym_infer:sym_infer -> ?⁠order:order -> ?⁠poly:poly -> ?⁠quants:bool -> warnings:(warning -> unit) -> file:Dolmen_std.Loc.file -> builtin_symbols -> env

Create a new environment.

Location helpers

val loc : env -> Dolmen_std.Loc.t -> Dolmen_std.Loc.full

Completes the location with the file name form the env.

val fragment_loc : env -> _ fragment -> Dolmen_std.Loc.full

Convenient function to get the location of a fragment.

val binding_reason : binding -> reason option

Extract the reason from a binding

raises Invalid_argument

if the binding is `Not_found

Name/Path helpers

val var_name : env -> Dolmen_std.Name.t -> string

Extract a variable name from a standard name.

val cst_path : env -> Dolmen_std.Name.t -> Dolmen_std.Path.t

Build a path from a standard name.

Bindings helpers

type var = [

| `Ty_var of ty_var

| `Term_var of term_var

| `Letin of env * Dolmen_std.Term.t * term_var * term

]

Variable bindings

type cst = [

| `Cstr of term_cstr

| `Field of term_field

| `Ty_cst of ty_cst

| `Term_cst of term_cst

]

Constant bindings

type builtin = [

| `Builtin of builtin_res

]

Builtin binding

type bound = [

| var

| cst

| builtin

]

val find_var : env -> Dolmen_std.Id.t -> [ var | not_found ]

Try and find the given id in the set of locally bound variables.

val find_global : env -> Dolmen_std.Id.t -> [ cst | not_found ]

Try and find the given id in the set of globally bound constants.

val find_builtin : env -> Dolmen_std.Id.t -> [ builtin | not_found ]

Try and find the given id in the set of bound builtin symbols.

val find_bound : env -> Dolmen_std.Id.t -> [ bound | not_found ]

Try and find a bound identifier in the env, whetehr it be locally bound (such as bound variables), constants bound at top-level, or builtin symbols bound by the builtin theory.

val get_global_custom : env -> 'a Dolmen_std.Tag.t -> 'a option

Get a custom value from the global environment.

val set_global_custom : env -> 'a Dolmen_std.Tag.t -> 'a -> unit

Set a custom value in the global environment.

Errors & Warnings

val _warn : env -> 'a fragment -> 'a warn -> unit

Emit a warning

val _error : env -> 'a fragment -> 'a err -> _

Raise an error

val suggest : limit:int -> env -> Dolmen_std.Id.t -> Dolmen_std.Id.t list

From a dolmen identifier, return a list of existing bound identifiers in the env that are up to ~limit in terms of distance of edition.

Parsing functions

val monomorphize : env -> Dolmen_std.Term.t -> term -> term

Monomorphize a term.

val parse_expr : res typer

Main parsing function.

val parse_ty : ty typer

val parse_term : term typer

val parse_prop : term typer

Wrappers around parse_expr to set the expect field of the env, and unwrap an expected return value.

val parse_app_ty_cst : (ty_cst -> Dolmen_std.Term.t list -> res) typer

val parse_app_term_cst : (term_cst -> Dolmen_std.Term.t list -> res) typer

Function used for parsing applications. The first dolmen term given is the application term being parsed (used for reporting errors).

val parse_app_ho_term : (term -> Dolmen_std.Term.t list -> res) typer

Function used for parsing an higher-order application.

val unwrap_ty : env -> Dolmen_std.Term.t -> res -> ty

val unwrap_term : env -> Dolmen_std.Term.t -> res -> term

Unwrap a result, raising the adequate typing error if the result if not as expected.

High-level functions

val decls : env -> ?⁠attrs:Dolmen_std.Term.t list -> Dolmen_std.Statement.decls -> [ `Type_decl of ty_cst | `Term_decl of term_cst ] list

Parse a list of potentially mutually recursive declarations.

val defs : env -> ?⁠attrs:Dolmen_std.Term.t list -> Dolmen_std.Statement.defs -> [ `Type_def of Dolmen_std.Id.t * ty_cst * ty_var list * ty | `Term_def of Dolmen_std.Id.t * term_cst * ty_var list * term_var list * term ] list

Parse a definition

val parse : term typer

Parse a formula