Parameter Tff.Ty

val bitv : int -> Type.Ty.t

Create a fixed size bitvector type.

val float : int -> int -> Type.Ty.t

Create a float type with fixed exponent size in bits and fixed significand, including the hidden bit.

val roundingMode : Type.Ty.t

Type of the rounding modes

type view = private [>

| `Real

| `Bitv of int

| `Float of int * int

]

Partial views for types. These are used in the Float theory to perform type-base dispatch for some conversion functions.

val view : Type.Ty.t -> view

Partial view of a type.