Parameter Tff.3-T

type t

The type of terms

type ty

The type of types.

val ty : t -> ty

Type of a term.

module Real : Dolmen_intf.Term.Smtlib_Float_Real with type t := t

Sub-module used for namespacing the real part of the theory requirements

module Bitv : Dolmen_intf.Term.Smtlib_Float_Bitv with type t := t

Sub-module used for namespacing the bitvector part of the theory requirements

module Float : Dolmen_intf.Term.Smtlib_Float_Float with type t := t

Sub-module used for namespacing the floating number part of the theory requirements