Module Dolmen_type.Intf

External Typechecker interface

This module defines the external typechcker interface, that is, the interface of an instantiated typechecker.

Typechecker interface

type symbol =

| Id of Dolmen.Std.Id.t

| Builtin of Dolmen.Std.Term.builtin

Wrapper around potential function symbols from the Dolmen AST.

module type Formulas = sig ... end

Typechecker interface