Module Smtlib2.Scan

This functor/module is there to help detect logic by scanning/inspecting types and terms

Parameters

module V : Dolmen_intf.View.TFF.S

Signature

exception Unknown_ty_builtin of V.ty_cst

exception Unknown_term_builtin of V.term_cst

Exception raised on builtins not allowed in SMT-LIB.

type acc

accumulator for computing minimal logics

val nothing : acc

The empty accumulator.

val need_univ : acc -> bool

Whether the univ type is used.

val need_unit : acc -> bool

Whether the unit type is used.

val to_logic : acc -> t

Tansform an accumulator into a logic

val scan_ty : acc -> V.Ty.t -> acc

Returns the minimal logic needed to typecheck a type.

val scan_term_decl : in_adt:bool -> acc -> V.Term.Cst.t -> acc

Returns the minimal logic needed to declare a term constant (i.e. including its type).

val scan_term : acc -> V.Term.t -> acc

Returns the minimal logic needed to typecheck a term.

val add_datatypes : acc -> acc

Add to the accumualtor the need to declare and use datatypes.

val add_free_sort : acc -> acc

Add to the accumualtor the need to declare new sorts/types.

val add_free_funs : acc -> acc

Add to the accumualtor the need to declare new functions. NOTE: declaring enw constants (i.e. functions with arity 0) is always allowed in all logics. This setting refers specifically to functions with an arity of at least 1.