Module Dolmen_type.Misc

Misc

Collection of various helper functions, some rather specific to dolmen (such as the bitv manipulation module), and others more generic (e.g. the

ist

and Option modules. For these generic functions, they are here either because they do not exist in the stdlib and they did not (yet) warrant depending on an external alternative stdlib, or because they were added in recent versions of ocaml and thus would unnecessarily restrict the availability of dolmen on older versions of ocaml.

module Options : sig ... end

Option helpers

module Lists : sig ... end

List helpers

module Strings : sig ... end

String helper

module Bitv : sig ... end

Bitvector helpers