Parameter Tff.3-T

include Dolmen_intf.Term.Smtlib_Arith_Common

type t

The type of terms

val mk : string -> t

Build a constant. The literal is passed as a string to avoid overflow caused by the limited precision of native number formats.

val minus : t -> t

Arithmetic unary minus/negation.

val add : t -> t -> t

Arithmetic addition.

val sub : t -> t -> t

Arithmetic substraction

val mul : t -> t -> t

Arithmetic multiplication

val lt : t -> t -> t

Arithmetic "less than" comparison.

val le : t -> t -> t

Arithmetic "less or equal" comparison.

val gt : t -> t -> t

Arithmetic "greater than" comparison.

val ge : t -> t -> t

Arithmetic "greater or equal" comparison.

val div : t -> t -> t

Euclidian division. See Smtlib theory for a full description.

val rem : t -> t -> t

Euclidiane integer remainder See Smtlib theory for a full description.

val abs : t -> t

Arithmetic absolute value.

val divisible : string -> t -> t

Arithmetic divisibility predicate. Indexed over constant integers (represented as strings, see int).