Module Ty.Const

A module for constant symbols the occur in types.

type t

The type of constant symbols the can occur in types

val compare : t -> t -> int

Comparison function on type constants.

val arity : t -> int

Return the arity of the given symbol.

val mk : path -> int -> t

Create a type constant with the given arity.

val set_tag : t -> 'a tag -> 'a -> unit

Set the value bound to a tag.

val add_tag : t -> 'a list tag -> 'a -> unit

Add a value to the list of values bound to a tag.