Module Ty.Var

A module for variables that occur in types.

type t

The type of variables the can occur in types

val compare : t -> t -> int

Comparison function on variables.

val mk : string -> t

Create a new type variable with the given name.