Module Dolmen_type__Misc.Fuzzy_Hashtbl

Fuzzy search Hashtables

type 'a t

The type of fuzzy hashtables from Dolmen identifiers to values of type 'a.

val create : unit -> _ t

Create a new fuzzy hashtable.

val find : 'a t -> Dolmen_std.Id.t -> 'a

Find the value bound to the given Identifier

raises Not_found

if the id is not bound.

val add : 'a t -> Dolmen_std.Id.t -> 'a -> unit

Add a new binding to the hashtable.

val suggest : 'a t -> limit:int -> Dolmen_std.Id.t -> Dolmen_std.Id.t list

Return a list of bound identifiers in the hashtbl that are close (up to limit in terms of edition distance) to the given id.