Module Dolmen_type__Tff

Typechecking of standard terms This module provides functions to typecheck terms from the untyped syntax tree defined in the standard implementation.

module type S = Dolmen_type.Tff_intf.S

Typechecker external interface

module Make : functor (Tag : Dolmen_intf.Tag.S) -> functor (Ty : Dolmen_intf.Ty.Tff with type 'a tag := 'a Tag.t and type path := Dolmen_std.Path.t) -> functor (T : Dolmen_intf.Term.Tff with type ty := Ty.t and type ty_var := Ty.Var.t and type ty_const := Ty.Const.t and type 'a tag := 'a Tag.t and type path := Dolmen_std.Path.t) -> S with module Tag = Tag and module Ty = Ty and module T = T