Parameter Tff.2-Ty

val int : Type.Ty.t

The type of integers, used as a default type of indexes when no type is provided

val array : Type.Ty.t -> Type.Ty.t -> Type.Ty.t

The type of functionnal arrays from one type to another.